filter graph fully working YAY.
