+++ src/components/Graph/Graph.jsx
@@ -1,8 +1,10 @@
 import { useHistory, useParams } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { useState, useEffect } from 'react';
-
 import { Line } from 'react-chartjs-2';
+
+// components
+import GraphForm from '../GraphForm/GraphForm';
 
 // material ui
 import Button from '@material-ui/core/Button';
@@ -23,18 +25,17 @@ function Graph() {
     // reducers
     const allGraph = useSelector(store => store.graphs.allGraph);
     const ropes = useSelector(store => store.formOptions.ropeReducer)
-	const walls = useSelector(store => store.formOptions.wallReducer)
-	const holds = useSelector(store => store.formOptions.holdReducer)
-    
+    const walls = useSelector(store => store.formOptions.wallReducer)
+    const holds = useSelector(store => store.formOptions.holdReducer)
+
 
     // local state
     const [open, setOpen] = useState(false);
     const [sendStatus, setSendStatus] = useState('error');
-    const [filterOptions, setFilterOptions] = useState('');
     const [rope, setRope] = useState('');
-	const [wall, setWall] = useState('');
-	const [hold, setHold] = useState('');
-	const [flash, setFlash] = useState('');
+    const [wall, setWall] = useState('');
+    const [hold, setHold] = useState('');
+    const [flash, setFlash] = useState('');
 
     // grabs all routes on page load
     useEffect(() => {
@@ -43,10 +44,10 @@ function Graph() {
             payload: {
                 gradeScheme: grading,
             },
-        },
-        {
-            type: 'FETCH_FORM_OPTIONS'
         })
+            // {
+            //     type: 'FETCH_FORM_OPTIONS'
+            // })
     }, []);
 
     // data for line graph. grabbed from reducer
@@ -88,85 +89,77 @@ function Graph() {
         history.push('/routes/list')
     }
 
-    // opens dialog form for sent/project selection
-    const handleFilter = () => {
-        setOpen(true);
-    }
-
-    // close dialog form without action
-    const handleFilterCancel = () => {
-        setOpen(false);
-    }
-
-    // on click of Filter button inside of form dialog, send info to server/db to grab selected routes
-    const handleFilterChoices = () => {
-        if (sendStatus == 'error') {
-            alert('Please choose what type of routes to see!');
-        } else {
-            dispatch({
-                type: 'FETCH_FILTERED_GRAPH',
-                payload: {
-                    gradeScheme: grading,
-                    sent: sendStatus,
-                    rope_type_id: 1,
-                    wall_id: 2,
-                    holds_id: 3,
-                    flash: true,
-                }
-            })
-            setOpen(false);
-            setFilterChip([
-                { key: 'true', label: 'sent' },
-                { key: 'false', label: 'project' },
-            ])
-        }
-    }
-
-    const useStyles = makeStyles((theme) => ({
-        root: {
-            display: 'flex',
-            justifyContent: 'center',
-            flexWrap: 'wrap',
-            listStyle: 'none',
-            padding: theme.spacing(0.5),
-            margin: 0,
-        },
-        chip: {
-            margin: theme.spacing(0.5),
-        },
-        // formControl: {
-        //     margin: theme.spacing(1),
-        //     minWidth: 120,
-        // },
-        // selectEmpty: {
-        //     marginTop: theme.spacing(2),
-        // },
-    }));
-
-    const classes = useStyles();
-
-    const [filterChip, setFilterChip] = useState([
-        { key: 'true', label: 'sent' },
-        { key: 'false', label: 'project' },
-        // { key: 'sendStatus', label: 'send status' },
-        // { key: 'rope', label: 'rope type' },
-        // { key: 'wall', label: 'wall angle' },
-        // { key: 'hold', label: 'main hold type' },
-        // { key: 'flash', label: 'if flashed' },
-        // { key: 'date', label: 'date' },
-    ]);
-
-    // const handleChipClick = (filter) => () => {
-    //     console.log('clicked chip');
-    //     setFilterOptions(filter.key);
-    //     // console.log('log chipToChoose', filter);
+    // // opens dialog form for sent/project selection
+    // const handleFilter = () => {
+    //     setOpen(true);
     // }
 
-    const handleChipClick = (chipToChoose) => () => {
-        setFilterChip((chips) => chips.filter((chip) => chip.label === chipToChoose.label));
-        // console.log('log sendStatusChip', sendStatus);
-        setSendStatus(chipToChoose.key);
-        console.log('log chipToChoose', chipToChoose);
+    // // close dialog form without action
+    // const handleFilterCancel = () => {
+    //     setOpen(false);
+    // }
+
+    // // on click of Filter button inside of form dialog, send info to server/db to grab selected routes
+    // const handleFilterChoices = () => {
+    //     if (sendStatus == 'error') {
+    //         alert('Please choose what type of routes to see!');
+    //     } else {
+    //         dispatch({
+    //             type: 'FETCH_FILTERED_GRAPH',
+    //             payload: {
+    //                 gradeScheme: grading,
+    //                 sent: sendStatus,
+    //                 rope_type_id: 1,
+    //                 wall_id: 2,
+    //                 holds_id: 3,
+    //                 flash: true,
+    //             }
+    //         })
+    //         setOpen(false);
+    //         setFilterChip([
+    //             { key: 'true', label: 'sent' },
+    //             { key: 'false', label: 'project' },
+    //         ])
+    //     }
+    // }
+
+    // const useStyles = makeStyles((theme) => ({
+    //     root: {
+    //         display: 'flex',
+    //         justifyContent: 'center',
+    //         flexWrap: 'wrap',
+    //         listStyle: 'none',
+    //         padding: theme.spacing(0.5),
+    //         margin: 0,
+    //     },
+    //     chip: {
+    //         margin: theme.spacing(0.5),
+    //     },
+    // }));
+
+    // const classes = useStyles();
+
+    // const [filterChip, setFilterChip] = useState([
+    //     { key: 'true', label: 'sent' },
+    //     { key: 'false', label: 'project' },
+    // ]);
+
+    // const handleChipClick = (chipToChoose) => () => {
+    //     setFilterChip((chips) => chips.filter((chip) => chip.label === chipToChoose.label));
+    //     // console.log('log sendStatusChip', sendStatus);
+    //     setSendStatus(chipToChoose.key);
+    //     console.log('log chipToChoose', chipToChoose);
+    // }
+
+    // const handleFilter = () => {
+    //     setOpen(true);
+    // }
+    // const handleFilterCancel = () => {
+    //     setOpen(false);
+    // }
+
+    const addOverlay = () => {
+        
     }
 
     return (
@@ -174,35 +167,8 @@ function Graph() {
             <h2>All Routes</h2>
             <Line data={data} options={options} />
             <Button onClick={handleBack}>Back</Button>
-            <Button onClick={handleFilter}>Filter Routes</Button>
-            <Dialog open={open} onClose={handleFilterCancel} aria-labelledby="form-dialog-title">
-                <DialogContent>
-                    <DialogContentText>
-                        Choose what type of routes to see.
-                    </DialogContentText>
-                    <div>
-                        {filterChip.map((data) => {
-                            return (
-                                <span key={data.key}>
-                                    <Chip
-                                        label={data.label}
-                                        onClick={handleChipClick(data)}
-                                        className={classes.chip}
-                                    />
-                                </span>
-                            );
-                        })}
-                    </div>
-                </DialogContent>
-                <DialogActions>
-                    <Button onClick={handleFilterCancel} color="primary">
-                        Cancel
-                    </Button>
-                    <Button onClick={handleFilterChoices} color="primary">
-                        Filter
-                    </Button>
-                </DialogActions>
-            </Dialog>
+            <Button onClick={addOverlay}>Overlay</Button>
+            <GraphForm/>
         </>
     )
 }
